backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
pypdf
sqlalchemy
python-dotenv
psycopg2-binary
